+++ kernel/include/minos/event.h
@@ -43,7 +43,7 @@ void event_task_wait(void *ev, int stat, uint32_t to);
 void __event_task_wait(unsigned long token, int mode, uint32_t to);
 void event_pend_down(struct task *task);
 
-int wait_event(void);
+long wait_event(void);
 
 static inline int event_has_waiter(struct event *ev)
 {
